requirements.txt
opencv-python==3.4.3.18
numpy==1.15.4
pandas==0.23.4
Keras==2.2.4
tensorflow-gpu==1.13.1
tqdm==4.29.1
scikit-learn==0.20.0
efficientnet==0.0.4
scikit-image==0.14.1
albumentations==0.3.0